pyproject.toml
[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.apis'

[project]
name = 'notifica'
version = '1.0'

authors = [
    {name = 'Yuri Teixeira', email = 'yuriteixeirac@proton.me'}
]

description = 'Uma RESTful API para suporte no sistema de comunicação do IFRN Campus Ceará-Mirim.'
readme = 'README.md'

requires-python = ">=3.12"

dependencies = [
    "django (>=6.0.3,<7.0.0)", 
    "mysqlclient (>=2.2.8,<3.0.0)", 
    "djangorestframework (>=3.17.0,<4.0.0)", 
    "djangorestframework-simplejwt[crypto] (>=5.5.1,<6.0.0)", 
    "requests (>=2.32.5,<3.0.0)", 
    "python-dotenv (>=1.2.2,<2.0.0)", 
    "cloudinary (>=1.44.1,<2.0.0)", 
    "google-genai (>=1.68.0,<2.0.0)"
]